% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbc_tsub.R
\name{calc_range}
\alias{calc_range}
\title{calculate temporal range recursively by overlap.}
\usage{
calc_range(v, min, max, res, overlap = 1)
}
\arguments{
\item{v}{sorted vector of temporal indicators}

\item{min}{cyclic minimum}

\item{max}{cyclic maximum}

\item{res}{resolution of time}

\item{overlap}{\link{integer}
The number of times to overlap the temporal indicators.}
}
\value{
a vector of temporal indicators
}
\description{
calculate temporal range recursively by overlap.
}
