% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_measures.R
\name{calc_wasserstein}
\alias{calc_wasserstein}
\title{Calculate Wasserstein Distance}
\usage{
calc_wasserstein(observed, model, n = nrow(observed), scale_dta = TRUE, ...)
}
\arguments{
\item{observed}{\link{data.frame}\cr
Measured (and interpolated) observations.}

\item{model}{\link{data.frame}\cr
Simulation climate data from the climate model.}

\item{n}{\code{integer(1)}\cr
Length of samples to be drawn from model and observed data.}

\item{scale_dta}{\code{logical(1)}\cr
If \code{TRUE} the data is scaled before the calculation of the Wasserstein.
Default is \code{TRUE}.}

\item{...}{\cr
Further arguments passed to \link[transport:transport]{transport::transport}.}
}
\value{
A named vector of length 2.\cr
\itemize{
\item Wasserstein_1: The Wasserstein distance of order 1
\item Wasserstein_2: The Wasserstein distance of order 2
}
}
\description{
Calculate the Wasserstein distance of order 1 and 2 between
multivariate distributions of model and observed data. Can be used to
evaluate the performance of a bias correction method by comparing the
distribution of the model data after the correction.
}
\examples{
#' \dontrun{
#' climate_sub = lapply(climate, function(x) x[, c("pr", "tas", "dew")])
#' margins_controls = list(xmin = c(0, NaN, NaN), type = c("zi", "c", "c"))
#' mp_vbc = vbc(climate_sub$oc, climate_sub$mc, climate_sub$mp,
#'              margins_controls = margins_controls, family_set = "tll",
#'              trunc_lvl = Inf)
#' class(mp_vbc)
#' summary(attr(mp_vbc, "vine_oc"))
#' summary(attr(mp_vbc, "vine_mp"))
#' calc_mci(climate_sub$mp, mp_vbc, time = climate$mp$time,
#'          margins_controls = margins_controls,
#'          family_set = "tll", trunc_lvl = Inf)
#' rbind(
#'  uncorrected = calc_wasserstein(climate_sub$oc[1:5000, ], climate_sub$mp[1:5000,]),
#'  corrected = calc_wasserstein(climate_sub$oc[1:5000, ], mp_vbc[1:5000, ])
#' )
#' }
}
\seealso{
\code{\link[transport:wasserstein]{transport::wasserstein()}}
}
