% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_margins.R
\name{calculate_margins}
\alias{calculate_margins}
\title{Calculate Margins}
\usage{
calculate_margins(dta, zero_inf, margins_controls, kde = NULL)
}
\arguments{
\item{dta}{Data to calculate the margins.}

\item{zero_inf}{\link{double}\cr
A vector of the length of climate variables to correct. The n-th element in
the vector indicates if the n-th climate variable is zero inflated.
\itemize{
\item \code{TRUE} indicates a zero inflated climate variable
\item \code{FALSE} indicates a non-zero inflated climate variable
Defaults to \code{FALSE} for all climate variables.
}}

\item{margins_controls}{\link{list}\cr
A list with arguments to be passed to \code{\link[kde1d:kde1d]{kde1d::kde1d()}}. Currently, there can be
\itemize{
\item \code{mult} numeric vector of length one or d; all bandwidths for marginal
kernel density estimation are multiplied with \code{mult}. Defaults to
\code{log(1 + d)} where \code{d} is the number of climate variables.
\item \code{xmin} numeric vector of length d; see \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
\item \code{xmax} numeric vector of length d; see \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
\item \code{bw} numeric vector of length d; see \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
\item \code{deg} numeric vector of length one or d; \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
}}

\item{kde}{A list of kernel density estimations.}
}
\value{
A data frame with the marginal PIT.
}
\description{
Calculate Margins
}
