% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_measures.R
\name{calculate_pcop}
\alias{calculate_pcop}
\title{Calculate Non-Exceedance Probability}
\usage{
calculate_pcop(data, margins_controls, ...)
}
\arguments{
\item{data}{\link{data.frame}\cr
Data to estimate the multivariate distribution.}

\item{margins_controls}{\link{list}\cr
A list with arguments to be passed to \code{\link[kde1d:kde1d]{kde1d::kde1d()}}. Currently, there can be
\itemize{
\item \code{mult} numeric vector of length one or d; all bandwidths for marginal
kernel density estimation are multiplied with \code{mult}. Defaults to
\code{log(1 + d)} where \code{d} is the number of climate variables.
\item \code{xmin} numeric vector of length d; see \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
\item \code{xmax} numeric vector of length d; see \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
\item \code{bw} numeric vector of length d; see \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
\item \code{deg} numeric vector of length one or d; \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
\item \code{type} character vector of length one or d; must be one of
c, cont, continuous for continuous variables, one of d, disc, discrete for
discrete integer variables, or one of zi, zinfl, zero-inflated for
zero-inflated variables.
}}

\item{...}{\cr
Arguments are passed to \link[rvinecopulib:vinecop]{rvinecopulib::vinecop} to specify the structure of
vines and margins. Note that the ellipsis of observed and model data are
specified with the same arguments.}
}
\value{
A data frame with the non-exceedance probability for the events.
}
\description{
Calculate Non-Exceedance Probability
}
