% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_measures.R
\name{calculate_pemp}
\alias{calculate_pemp}
\title{Calculate Non-Exceedance Probability}
\usage{
calculate_pemp(data)
}
\arguments{
\item{data}{\link{data.table}\cr
On each of the rows in the data.table, the non-exceedance probability is
calculated.}
}
\value{
A data frame with the non-exceedance probability for the events. The
non-exceedance probability is the percentage of rows that satisfy the
condition that all values in the row are less or equal to the values in the
row.
}
\description{
Calculate Non-Exceedance Probability
}
