% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vine_correct.R
\name{check_vbc_args}
\alias{check_vbc_args}
\title{Check arguments for vine correction}
\usage{
check_vbc_args(oc, mc, mp, var_names, digits, zero_inf)
}
\arguments{
\item{oc}{\link{data.frame}\cr
Measured (and interpolated) observations during calibration period.}

\item{mc}{\link{data.frame}\cr
Simulation data from a climate model during calibration period.}

\item{mp}{\link{data.frame}\cr
Simulation data from a climate model during projection period.}

\item{var_names}{\link{character}\cr
Names of corrected climate data. Defaults to the column names of the
observed data \code{oc}.}

\item{digits}{\code{integer(1)}\cr
The corrected data will be rounded to the specified digit length after
correction. Defaults to \code{NA} where no rounding is performed.}

\item{zero_inf}{\link{double}\cr
A vector of the length of climate variables to correct. The n-th element in
the vector indicates if the n-th climate variable is zero inflated.
\itemize{
\item \code{TRUE} indicates a zero inflated climate variable
\item \code{FALSE} indicates a non-zero inflated climate variable
Defaults to \code{FALSE} for all climate variables.
}}
}
\description{
Check arguments for vine correction
}
