% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.R
\docType{data}
\name{climate}
\alias{climate}
\title{Sample CRCM5 and SDCLIREF data}
\format{
A list with four elements
\describe{
\item{oc}{Measured observations during calibration period.}
\item{op}{Measured observations during projection period.}
\item{mc}{Simulation data from the model during calibration period.}
\item{mp}{Simulation data from the model during projection period.}
}
}
\usage{
climate
}
\description{
Sample CRCM5-LE member "kba" and SDCLIREF v1 data for the city of Munich
(48°8'14.75"N, 11°34'31.76"E).
}
\details{
Data are available in 3 hourly resolution from:

01.01.1991 - 31.12.2000 for calibration

01.01.2001 - 31.12-2010 for projection.

With 5 climate variables to correct:

tas: average surface temperature (deg. C)

pr: precipitation (mm day-1)

dew: dewpoint temperature (deg. C)

rsds: surface downwelling shortwave radiation (W m-2)

sfcWind: surface wind speed (m s-1)

time: date and time in three hourly resolution
}
\keyword{datasets}
