% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_data.R
\name{correct_rosenblatt}
\alias{correct_rosenblatt}
\title{Correct data by Rosenblatt transformation}
\usage{
correct_rosenblatt(mpu, ocu, z_inf = FALSE)
}
\arguments{
\item{mpu}{\link{data.table}\cr
Model data from \code{\link[=vine_correct]{vine_correct()}}.}

\item{ocu}{\link{data.table}\cr
Observed data from \code{\link[=vine_correct]{vine_correct()}}.}

\item{z_inf}{\code{logical(1)}\cr
If \code{TRUE} at least one margin is zero inflated.}
}
\value{
Corrected data by Rosenblatt transformation.
}
\description{
Correct data by Rosenblatt transformation
}
