% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_margins.R
\name{estimate_margins}
\alias{estimate_margins}
\title{Estimate Margins}
\usage{
estimate_margins(dta, margins_controls, kde = NULL)
}
\arguments{
\item{dta}{\link[data.table:data.table]{data.table::data.table}
Data to calculate the margins.}

\item{margins_controls}{\link{list}
A list with arguments to be passed to \code{\link[kde1d:kde1d]{kde1d::kde1d()}}. Currently, there can be
\itemize{
\item \code{mult} numeric vector of length one or d; all bandwidths for marginal
kernel density estimation are multiplied with \code{mult}. Defaults to
\code{log(1 + d)} where \code{d} is the number of climate variables.
\item \code{xmin} numeric vector of length d; see \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
\item \code{xmax} numeric vector of length d; see \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
\item \code{bw} numeric vector of length d; see \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
\item \code{deg} numeric vector of length one or d; \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
\item \code{type} character vector of length one or d; must be one of
c, cont, continuous for continuous variables, one of d, disc, discrete for
discrete integer variables, or one of zi, zinfl, zero-inflated for
zero-inflated variables.
}}

\item{kde}{\link{list}
kernel density estimations.}
}
\value{
A data frame with the marginal PIT.
}
\description{
Estimate Margins
}
