% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_margins.R
\name{expand_margin_controls}
\alias{expand_margin_controls}
\title{Expand margins}
\usage{
expand_margin_controls(controls, dta)
}
\arguments{
\item{controls}{margins_controls}

\item{dta}{\link[data.table:data.table]{data.table::data.table}\cr
Data to calculate the margins.}
}
\value{
Expanded margins.
}
\description{
Expand margins
}
