% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vine_correct.R
\name{flatten_zi_margins}
\alias{flatten_zi_margins}
\title{Flatten zero inflated margins}
\usage{
flatten_zi_margins(margins, zero_inf, eps = 1e-10)
}
\arguments{
\item{margins}{\link{data.frame}\cr
The margins of the climate data.}

\item{zero_inf}{\link{double}\cr
A vector of the length of climate variables to correct. The n-th element in
the vector indicates if the n-th climate variable is zero inflated.
\itemize{
\item \code{TRUE} indicates a zero inflated climate variable
\item \code{FALSE} indicates a non-zero inflated climate variable
Defaults to \code{FALSE} for all climate variables.
}}

\item{eps}{\link{double}\cr
A small number to avoid zero and one values in zero inflated margins.}
}
\value{
Flatten zero inflated margins.
}
\description{
Flatten zero inflated margins
}
