% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbc_tsub.R
\name{invert_list}
\alias{invert_list}
\title{Invert a list}
\usage{
invert_list(list)
}
\arguments{
\item{list}{\link{list}\cr
A list of depth 2. The first depth indicates the temporal subset and the
second depth the members of the model ensemble.}
}
\value{
\link{list}\cr
A list of depth 2. The first depth indicates the members of the model
ensemble and the second depth the temporal subset.
}
\description{
Invert a list
}
