% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data.R
\name{model_vine}
\alias{model_vine}
\title{Model Multivariate Distribution by Vine Copula}
\usage{
model_vine(data, margins_controls, ...)
}
\arguments{
\item{data}{\link{data.frame}
Data to estimate the multivariate distribution.}

\item{margins_controls}{\link{list}
A list with arguments to be passed to \code{\link[kde1d:kde1d]{kde1d::kde1d()}}. Currently, there can be
\itemize{
\item \code{mult} numeric vector of length one or d; all bandwidths for marginal
kernel density estimation are multiplied with \code{mult}. Defaults to
\code{log(1 + d)} where \code{d} is the number of climate variables.
\item \code{xmin} numeric vector of length d; see \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
\item \code{xmax} numeric vector of length d; see \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
\item \code{bw} numeric vector of length d; see \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
\item \code{deg} numeric vector of length one or d; \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
\item \code{type} character vector of length one or d; must be one of
c, cont, continuous for continuous variables, one of d, disc, discrete for
discrete integer variables, or one of zi, zinfl, zero-inflated for
zero-inflated variables.
}}

\item{...}{Arguments are passed to \link[rvinecopulib:vinecop]{rvinecopulib::vinecop} to specify the structure of
vines and margins. Note that the ellipsis of observed and model data are
specified with the same arguments.}
}
\value{
The PIT-transformed margins from \code{\link[=estimate_margins]{estimate_margins()}}. Additionally
the data frame contains the attribute \code{vine} with the vine copula model and
the attribute \code{kde} with the kernel density estimation of the data.
}
\description{
Estimate the multivariate distribution of the model data via vine copula
estimation (see \link[rvinecopulib:vine]{rvinecopulib::vine}).
}
