% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tails.R
\name{plot_tails}
\alias{plot_tails}
\title{Plot Tails}
\usage{
plot_tails(data, var, xmin = NA, scale_d = 3, offset = 8, length_out = 10000)
}
\arguments{
\item{data}{\link{data.table}\cr
A data.table containing the data to be plotted.}

\item{var}{\link{character}\cr
The variable to be plotted. Must be a column of \code{data}.}

\item{xmin}{\link{numeric}\cr
The minimum value of the variable to be plotted. Passed to \link{kde1d}.
Default is \code{NA} indicating no lower limit.}

\item{scale_d}{\link{numeric}\cr
The scaling factor for the density plot between continuous and relative axis.
Default is \code{3}. Regulate the factor, if continuous density estimates and
relative frequency are too distinct on the y-axis. Only used for zero
inflated variables.}

\item{offset}{\link{numeric}\cr
The offset to be used for the log transformation. Default is \code{8}.
Reduce the offset towards 0 if bar and kde curve if the gap between zero
inflation bar and continuous kde curve is too large. Only used
for zero inflated variables.}

\item{length_out}{\link{numeric}\cr
The length of the sequence to be used for the density plot. Default is \code{1e4}.}
}
\value{
A ggplot that plots the continuous density estimate of the variable
and the relative frequency of inflation.
}
\description{
This function plots the tails probabilities of a marginal
distribution.
}
