% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosenblatt.R
\name{rosenblatt}
\alias{rosenblatt}
\title{Rosenblatt transformation}
\usage{
rosenblatt(x, model, cores = 1, randomize_discrete = TRUE)
}
\description{
The Rosenblatt transformation is a method to transform uniform
pseudo-observations into the original scale of the data, preserving the
copula structure. Adapted from \code{\link[rvinecopulib:rosenblatt]{rvinecopulib::rosenblatt()}} and
\code{\link[rvinecopulib:rosenblatt]{rvinecopulib::inverse_rosenblatt()}}.
}
