% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbc_tsub.R
\name{subset_time}
\alias{subset_time}
\title{subset time and final reading index}
\usage{
subset_time(data, hrs, mnt, overlap = 1)
}
\arguments{
\item{data}{\link{data.frame}\cr
Data to estimate the multivariate distribution.}

\item{hrs}{\link{integer}\cr
The hours of the day to be used for subsetting.}

\item{mnt}{\link{integer}\cr
The months of the year to be used for subsetting.}

\item{overlap}{\link{integer}\cr
The number of times to overlap the temporal indicators.}
}
\description{
subset time and final reading index
}
