% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbc.R
\name{vbc}
\alias{vbc}
\title{Correct zero inflated climate model simulations via vine copulas}
\usage{
vbc(
  oc,
  mc,
  mp,
  var_names = colnames(oc),
  margins_controls = list(mult = NULL, xmin = NaN, xmax = NaN, bw = NA, deg = 2, type =
    "c"),
  time_mp = NA,
  ...
)
}
\arguments{
\item{oc}{\link{data.table}\cr
Measured (and interpolated) observations during calibration period.}

\item{mc}{\link{data.table}\cr
Simulation data from a climate model during calibration period.}

\item{mp}{\link{data.table}\cr
Simulation data from a climate model during projection period.}

\item{var_names}{\link{character}\cr
Names of corrected climate data. Defaults to the column names of the
observed data \code{oc}.}

\item{margins_controls}{\link{list}\cr
A list with arguments to be passed to \code{\link[kde1d:kde1d]{kde1d::kde1d()}}. Currently, there can be
\itemize{
\item \code{mult} numeric vector of length one or d; all bandwidths for marginal
kernel density estimation are multiplied with \code{mult}. Defaults to
\code{log(1 + d)} where \code{d} is the number of climate variables.
\item \code{xmin} numeric vector of length d; see \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
\item \code{xmax} numeric vector of length d; see \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
\item \code{bw} numeric vector of length d; see \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
\item \code{deg} numeric vector of length one or d; \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
\item \code{type} character vector of length one or d; must be one of
c, cont, continuous for continuous variables, one of d, disc, discrete for
discrete integer variables, or one of zi, zinfl, zero-inflated for
zero-inflated variables.
}}

\item{time_mp}{\link{numeric}\cr
Time vector of the projection period attached to the returned object as a
column. Defaults to \code{NA}, which means no time vector is attached.}

\item{...}{\cr
Arguments are passed to \link[rvinecopulib:vinecop]{rvinecopulib::vinecop} to specify the structure of
vines and margins. Note that the ellipsis of observed and model data are
specified with the same arguments.}
}
\value{
\link{data.table}\cr
The corrected projection period data in \code{mp}. Additionally the data frame
contains the attributes \code{vine_oc}, \code{kde_oc}, \code{vine_mp}, and \code{kde_mp} which
store the vine copula and kernel density estimation objects of the observed
and model data.
}
\description{
The multivariate distribution functions of model and observed are estimated
via vine copula estimation (see \link[rvinecopulib:vine]{rvinecopulib::vine}).
This method provides the tool to model the multivariate distribution function
of the climate data by Vine Copulas. The climate variables may be zero
inflated.
The quantiles are then mapped via (inverse) Rosenblatt transformation from
the model to the observed distribution. The corrected data are mapped to the
projection domain via delta mapping. The steps are equivalent to those in
univariate bias correction via quantile delta mapping.
}
\examples{
#' \dontrun{
#' data("climate")
#' climate_sub = lapply(climate, function(x) x[, c("pr", "tas", "dew", "time")])
#' margins_controls = list(xmin = c(0, NaN, NaN), type = c("zi", "c", "c"))
#' t_subs = list(day = list(hours = c(6, 9, 12, 15), month = 1:12),
#'               night = list(hours = c(18, 21, 0, 3), month = 1:12))
#' mp_vbc = vbc_tsub(climate_sub$oc, climate_sub$mc, climate_sub$mp,
#'                   t_subs = t_subs, margins_controls = margins_controls,
#'                   family_set = "tll", trunc_lvl = Inf)
#' class(mp_vbc)
#' summary(attr(mp_vbc, "vine_oc"))
#' summary(attr(mp_vbc, "vine_mp"))
#' calc_mci(climate_sub$mp, mp_vbc, time = climate$mp$time,
#'          margins_controls = margins_controls,
#'          family_set = "tll", trunc_lvl = Inf)
#' rbind(
#'  uncorrected = calc_wasserstein(climate_sub$oc[1:5000, ], climate_sub$mp[1:5000,]),
#'  corrected = calc_wasserstein(climate_sub$oc[1:5000, ], mp_vbc[1:5000, ])
#' )
#' }
}
\references{
Czado, C. (2019). Analyzing dependent data with vine copulas. Lecture Notes
in Statistics, Springer, 222.

Rosenblatt, M. (1952). Remarks on a multivariate transformation. The annals
of mathematical statistics, 23(3), 470-472.

Cannon, A.J., S.R. Sobie, and T.Q. Murdock, (2015). Bias correction of
simulated precipitation by quantile mapping: How well do methods preserve
relative changes in quantiles and extremes? Journal of Climate, 28:6938-6959.
}
