% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbc.R
\name{vbc_ensemble}
\alias{vbc_ensemble}
\title{Correction by VBC for model ensembles}
\usage{
vbc_ensemble(mp, mc, rc, var_names, margins_controls, time_mp, ...)
}
\arguments{
\item{mp}{\link[data.table:data.table]{data.table::data.table} OR \link{list}\cr
Simulation data from a climate model during projection period. If this is a
list, the algorithm expects each element to be a member of a model ensemble.
Each list element is then a data table.}

\item{mc}{\link[data.table:data.table]{data.table::data.table}\cr
Simulation data from a climate model during calibration period.}

\item{rc}{\link[data.table:data.table]{data.table::data.table}\cr
Historical reference in calibration period.}

\item{var_names}{\link{character}\cr
Names of corrected climate data. Defaults to the column names of the
observed data \code{rc}.}

\item{margins_controls}{\link{list}\cr
A list with arguments to be passed to \code{\link[kde1d:kde1d]{kde1d::kde1d()}}. Currently, there can be
\itemize{
\item \code{mult} numeric vector of length one or d; all bandwidths for marginal
kernel density estimation are multiplied with \code{mult}. Defaults to
\code{log(1 + d)} where \code{d} is the number of climate variables.
\item \code{xmin} numeric vector of length d; see \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
\item \code{xmax} numeric vector of length d; see \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
\item \code{bw} numeric vector of length d; see \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
\item \code{deg} numeric vector of length one or d; \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.
\item \code{type} character vector of length one or d; must be one of
c, cont, continuous for continuous variables, one of d, disc, discrete for
discrete integer variables, or one of zi, zinfl, zero-inflated for
zero-inflated variables.
}}

\item{time_mp}{\link{numeric}\cr
Time vector of the projection period attached to the returned object as a
column. Defaults to \code{NA}, which means no time vector is attached.}

\item{...}{\cr
Arguments are passed to \link[rvinecopulib:vinecop]{rvinecopulib::vinecop} to specify the structure of
vines and margins. Note that the ellipsis of observed and model data are
specified with the same arguments.}
}
\value{
\link{list}\cr
A list of corrected data frames for each member of the model ensemble.
}
\description{
Correction by VBC for model ensembles
}
