% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_measures.R
\name{calc_wasserstein}
\alias{calc_wasserstein}
\title{Calculate Wasserstein Distance}
\usage{
calc_wasserstein(observed, model, n = nrow(observed), scale_dta = TRUE, ...)
}
\arguments{
\item{observed}{\link{data.frame}\cr
Measured (and interpolated) observations.}

\item{model}{\link{data.frame}\cr
Simulation climate data from the climate model.}

\item{n}{\code{integer(1)}\cr
Length of samples to be drawn from model and observed data.}

\item{scale_dta}{\code{logical(1)}\cr
If \code{TRUE} the data is scaled before the calculation of the Wasserstein.
Default is \code{TRUE}.}

\item{...}{\cr
Further arguments passed to \link[transport:transport]{transport::transport}.}
}
\value{
A named vector of length 2.\cr
\itemize{
\item Wasserstein_1: The Wasserstein distance of order 1
\item Wasserstein_2: The Wasserstein distance of order 2
}
}
\description{
Calculate the Wasserstein distance of order 1 and 2 between
multivariate distributions of model and observed data. Can be used to
evaluate the performance of a bias correction method by comparing the
distribution of the model data after the correction.
}
\examples{
#' \dontrun{
#' library(MBC)
#' data("climate")
#' climate_sub = lapply(climate, function(x) x[, c("pr", "tas", "dew")])
#' margins_controls = list(xmin = c(0, NaN, NaN))
#' zero_inf = c(TRUE, FALSE, FALSE)
#' mp_vbc = vine_correct(climate_sub$oc, climate_sub$mc, climate_sub$mp,
#'                       margins_controls = margins_controls,
#'                       zero_inf = zero_inf, family_set = "tll",
#'                       trunc_lvl = Inf)
#' summary(attr(mp_vbc, "vine_oc"))
#' summary(attr(mp_vbc, "vine_mp"))
#' mp_ubc = mapply(QDM, o.c = climate_sub$oc, m.c = climate_sub$mc,
#'                 m.p = climate_sub$mp, ratio = c(TRUE, FALSE, FALSE),
#'                 trace = c(0.01, Inf, Inf), SIMPLIFY = FALSE)
#' mp_ubc = do.call(cbind.data.frame, lapply(mp_ubc, function(var) var$mhat.p))
#' mp_mbcn = data.frame(MBCn(o.c = climate_sub$oc, m.c = climate_sub$mc,
#'                           m.p = climate_sub$mp, ratio.seq = c(TRUE, FALSE, FALSE),
#'                           trace = c(0.01, Inf, Inf), silent = TRUE)$mhat.p)
#' colnames(mp_mbcn) = colnames(mp_vbc)
#' c(
#'   "ubc" = attr(calc_mci(climate_sub$mp, mp_ubc,
#'                         margins_controls = margins_controls,
#'                         zero_inf = zero_inf, family_set = "tll",
#'                         trunc_lvl = Inf), "global"),
#'   "vbc" = attr(calc_mci(climate_sub$mp, mp_vbc,
#'                         margins_controls = margins_controls,
#'                         zero_inf = zero_inf, family_set = "tll",
#'                         trunc_lvl = Inf), "global"),
#'   "mbcn" = attr(calc_mci(climate_sub$mp, mp_mbcn,
#'                          margins_controls = margins_controls,
#'                          zero_inf = zero_inf, family_set = "tll",
#'                          trunc_lvl = Inf), "global")
#' )
#'
#' list(
#'   "ubc" = calc_wasserstein(climate_sub$oc, mp_ubc),
#'   "vbc" = calc_wasserstein(climate_sub$oc, mp_vbc),
#'   "mbcn" = calc_wasserstein(climate_sub$oc, mp_mbcn)
#' )
#' }
}
\seealso{
\code{\link[transport:wasserstein]{transport::wasserstein()}}
}
