% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbc.R
\name{check_vbc_args}
\alias{check_vbc_args}
\title{Check arguments for vine correction}
\usage{
check_vbc_args(oc, mc, mp, var_names)
}
\arguments{
\item{oc}{\link{data.table}\cr
Measured (and interpolated) observations during calibration period.}

\item{mc}{\link{data.table}\cr
Simulation data from a climate model during calibration period.}

\item{mp}{\link{data.table}\cr
Simulation data from a climate model during projection period.}

\item{var_names}{\link{character}\cr
Names of corrected climate data. Defaults to the column names of the
observed data \code{oc}.}
}
\description{
Check arguments for vine correction
}
