% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbc.R
\name{check_vbc_args}
\alias{check_vbc_args}
\title{Check arguments for vine correction}
\usage{
check_vbc_args(mp, mc, rc, var_names)
}
\arguments{
\item{mp}{\link{data.table} OR \link{list}\cr
Simulation data from a climate model during projection period. If this is a
list, the algorithm expects each element to be a member of a model ensemble.
Each list element is then a data table.}

\item{mc}{\link{data.table}\cr
Simulation data from a climate model during calibration period.}

\item{rc}{\link{data.table}\cr
Historical reference in calibration period.}

\item{var_names}{\link{character}\cr
Names of corrected climate data. Defaults to the column names of the
observed data \code{rc}.}
}
\description{
Check arguments for vine correction
}
