% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_data.R
\name{correct_rosenblatt}
\alias{correct_rosenblatt}
\title{Correct data by Rosenblatt transformation}
\usage{
correct_rosenblatt(mpu, rcu)
}
\arguments{
\item{mpu}{\link[data.table:data.table]{data.table::data.table}\cr
Model data from \code{\link[=vbc]{vbc()}}.}

\item{rcu}{\link[data.table:data.table]{data.table::data.table}\cr
Observed data from \code{\link[=vbc]{vbc()}}.}
}
\value{
Corrected data by Rosenblatt transformation.
}
\description{
Correct data by Rosenblatt transformation
}
